requirements.txt
streamlit==1.44.1
Pillow==11.1.0
geopy==2.4.1
tensorflow==2.16.1
numpy>=1.26.0,<=1.26.4
streamlit-option-menu==0.3.2
streamlit-aggrid==0.2.2.post4
streamlit-geolocation==0.0.10
